Better organize new unified node definition folder
