+++ inst/editor/src/components/Tabs/Tabset/Tab.tsx
@@ -4,12 +4,12 @@ import { useSelector } from "react-redux";
 
 import { useCurrentSelection } from "../../../state/selectedPath";
 import type { RootState } from "../../../state/store";
-import { isKnownShinyUiNode } from "../../../ui-node-definitions/isShinyUiNode";
 import type { NodePath } from "../../../ui-node-definitions/NodePath";
-import { makeChildPath } from "../../../ui-node-definitions/nodePathUtils";
 import type { ShinyUiNode } from "../../../ui-node-definitions/ShinyUiNode";
 import { getNode } from "../../../ui-node-definitions/TreeManipulation/getNode";
 import { samePath } from "../../../ui-node-definitions/TreeManipulation/samePath";
+import { isKnownShinyUiNode } from "../../../ui-node-definitions/utils/isShinyUiNode";
+import { makeChildPath } from "../../../ui-node-definitions/utils/nodePathUtils";
 import { NodeWrapper } from "../../UiNode/NodeWraper";
 
 import classes from "./Tabset.module.css";
