--- inst/editor/src/ui-node-definitions/index.ts
@@ -1,4 +1,0 @@
-import type { node_info_by_id } from "./uiNodeTypes";
-
-export { node_info_by_id } from "./uiNodeTypes";
-export type NodeInfoById = typeof node_info_by_id;
